+++ packages/design-tokens/src/system/color/specifiable.ts
@@ -60,12 +60,6 @@ export interface IntentColors {
    */
   link: string
   /**
-   * Link text color on interaction
-   * Used for: Link :active, :focused and :hover states
-   * @default blue700
-   */
-  linkInteractive: string
-  /**
    * Critical intent color
    * Used for: Delete button, error and validation messages
    * @default red500
@@ -96,4 +90,13 @@ export interface IntentColors {
    * @default blue500
    */
   inform: string
+}
+
+export interface IntentDerivativeColors {
+  /**
+   * Link text color on interaction
+   * Used for: Link :active, :focused and :hover states
+   * @default blue700
+   */
+  linkInteractive: string
 }