+++ packages/design-tokens/src/tokens/color/index.ts
@@ -27,7 +27,6 @@
 import {
   blue500,
   blue600,
-  blue700,
   charcoal500,
   charcoal800,
   green500,
@@ -37,9 +36,8 @@ import {
   white,
 } from '../../legacy/palette'
 import { IntentColors } from '../../system/color/specifiable'
-import { CoreColors, Colors, SpecifiableColors } from '../../system'
-import { generateSurfaceColors } from '../../utils/color/surface'
-import { fallbackBlends, fallbackStateful } from './fallbacks'
+import { CoreColors } from '../../system'
+import { generateColors } from '../../utils/color'
 
 /* eslint-disable sort-keys-fix/sort-keys-fix */
 
@@ -51,7 +49,6 @@ export const defaultCoreColors: CoreColors = {
 
 export const defaultIntentColors: IntentColors = {
   link: blue600,
-  linkInteractive: blue700,
   critical: red500,
   warn: yellow500,
   neutral: charcoal500,
@@ -59,14 +56,7 @@ export const defaultIntentColors: IntentColors = {
   inform: blue500,
 }
 
-const specifiableColors: SpecifiableColors = {
+export const colors = generateColors({
   ...defaultCoreColors,
-  ...generateSurfaceColors(defaultCoreColors),
   ...defaultIntentColors,
-}
-
-export const colors: Colors = {
-  ...specifiableColors,
-  ...fallbackBlends,
-  ...fallbackStateful,
-}
+})