+++ packages/design-tokens/src/tokens/typography/font_families.ts
@@ -25,8 +25,9 @@
  */
 
 import { FontFamilyChoices } from '../../system'
+import { generateFontFamilies } from '../../utils/typography'
 
-const basicFallbacks = [
+const sansSerifFontFallbacks = [
   '"Noto Sans JP"',
   '"Noto Sans CJK KR"',
   '"Noto Sans Arabic UI"',
@@ -38,9 +39,9 @@ const basicFallbacks = [
   'sans-serif',
 ]
 
-const defaultFallbacks = {
-  body: basicFallbacks,
-  brand: basicFallbacks,
+export const defaultFontFallbacks = {
+  body: sansSerifFontFallbacks,
+  brand: sansSerifFontFallbacks,
   code: [
     '"Monaco"',
     '"Menlo"',
@@ -51,40 +52,10 @@ const defaultFallbacks = {
   ],
 }
 
-export const fontFacesToFamily = (
-  faces: string[] | string,
-  fallbacks: string[]
-) => {
-  if (typeof faces === 'string') {
-    faces = [faces]
-  }
-
-  faces = [...faces, ...fallbacks]
-
-  return faces.map((face) => `${face}`).join(', ')
-}
-
 export const defaultFonts: FontFamilyChoices = {
   body: '"Roboto"',
   brand: '"Red Hat Display"',
   code: '"Roboto Mono"',
 }
 
-export const constructFontStack = (fontStack: Partial<FontFamilyChoices>) => {
-  Object.keys(fontStack).forEach((key) =>
-    !fontStack[key] ? delete fontStack[key] : {}
-  )
-
-  const fontFamilies: FontFamilyChoices = {
-    ...defaultFonts,
-    ...fontStack,
-  }
-
-  Object.entries(fontFamilies).map(([key, fontFace]) => {
-    fontFamilies[key] = fontFacesToFamily(fontFace, defaultFallbacks[key])
-  })
-
-  return fontFamilies
-}
-
-export const fontFamilies = constructFontStack(defaultFonts)
+export const fontFamilies = generateFontFamilies(defaultFonts)