+++ packages/design-tokens/src/utils/color/intent.ts
@@ -24,37 +24,11 @@
 
  */
 
-import pickBy from 'lodash/pickBy'
-import identity from 'lodash/identity'
-import {
-  Colors,
-  CoreColors,
-  IntentColors,
-  SpecifiableColors,
-} from '../../system'
-import { generateSurfaceColors } from './surface'
-import { generateBlendColors } from './blend'
-import { generateStatefulColors } from './stateful'
-import { generateIntentDerivatives } from './intent'
+import { IntentColors } from '../../system'
+import { generateInteractive } from './stateful'
 
-export const generateColors = (
-  themeColors: CoreColors & IntentColors,
-  customColors?: Partial<CoreColors & IntentColors>
-): Colors => {
-  const colors = { ...themeColors, ...pickBy(customColors, identity) }
-
-  const specifiableColors: SpecifiableColors = {
-    ...colors,
-    ...generateSurfaceColors(colors),
-    ...generateIntentDerivatives(colors),
-  }
-
-  const blends = generateBlendColors(specifiableColors)
-  const statefulColors = generateStatefulColors(specifiableColors)
-
+export const generateIntentDerivatives = (intentColors: IntentColors) => {
   return {
-    ...specifiableColors,
-    ...blends,
-    ...statefulColors,
+    linkInteractive: generateInteractive(intentColors.link),
   }
 }